changed code search,desktop added vfs.js and dir search
